DEV: Convert `badge-preview` modal to component-based API
